# Conflicts:
#   app/Http/Controllers/EventController.php
#   app/Http/Controllers/UserController.php
#   app/Policies/EventPolicy.php
#   routes/web.php

+++ app/Http/Controllers/UserController.php
@@ -4,30 +4,24 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
-
-    /**
-     * Display the specified resource.
-     */
     public function show(string $id)
     {
         $user = User::findOrFail($id);
-        $events = $user->ownedEvents();
+        
         $this->authorize('show', $user);
+        
         return view('pages.user', [
             'user' => $user
         ]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit(string $id)
     {
         $user = User::findOrFail($id);
+
         $this->authorize('edit', $user);
 
         return view('pages.user_edit', [
@@ -36,15 +30,12 @@ class UserController extends Controller
     }
 
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, string $id)
     {
-
         $user = User::findOrFail($id);
+        
         $this->authorize('update', $user);
-    
+
         $request->validate([
             'email' => 'required|email|max:250|unique:users,email,' . $id,
             'username' => 'required|string|max:250|unique:users,username,' . $id,
@@ -52,30 +43,28 @@ class UserController extends Controller
             'description' => 'string|max:1000',
             'password' => 'nullable|min:8',
         ]);
-    
+
         if ($request->filled('password')) {
             $user->password = bcrypt($request->password);
         }
-    
+
         $user->email = $request->email;
         $user->username = $request->username;
         $user->name = $request->name;
         $user->description = $request->description;
-    
+
         $user->save();
-    
+
         return redirect()->route('user.show', ['id' => $user->id])
             ->withSuccess('You have successfully edited your profile!');
     }
-    
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function delete(string $id) //fazer com que os eventos/comentarios/etc nao sejam apagados
     {
         $user = User::findOrFail($id);
-        //authorize delete
+
+        $this->authorize('delete', $user);
+
         $user->delete();
         return redirect()->route('home')
             ->withSuccess('You have successfully deleted your profile!');
